:lipstick: Create Parse backend module
