+++ accountAuthenticator/src/main/java/com/udinic/accounts_authenticator_example/authentication/backend/AccountGeneral.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017 Udi Cohen
+ * Copyright (c) 2017 Udi Cohen, Joao Paulo Fernandes Ventura
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.udinic.accounts_authenticator_example.authentication;
+package com.udinic.accounts_authenticator_example.authentication.backend;
 
 public class AccountGeneral {
