+++ accountAuthenticator/src/main/java/com/udinic/accounts_authenticator_example/authentication/backend/IParseEndpoint.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017 Udi Cohen
+ * Copyright (c) 2017 Udi Cohen, Joao Paulo Fernandes Ventura
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,18 +14,12 @@
  * limitations under the License.
  */
 
-package com.udinic.accounts_authenticator_example.authentication;
+package com.udinic.accounts_authenticator_example.authentication.backend;
 
-import android.app.Service;
-import android.content.Intent;
-import android.os.IBinder;
+public interface IParseEndpoint {
 
-public class UdinicAuthenticatorService extends Service {
+    String userSignUp(final String name, final String email, final String password) throws Exception;
 
-    @Override
-    public IBinder onBind(Intent intent) {
-        UdinicAuthenticator authenticator = new UdinicAuthenticator(this);
-        return authenticator.getIBinder();
-    }
+    String userSignIn(final String user, final String pass) throws Exception;
 
 }
